requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
onnxruntime==1.19.0
Pillow==10.4.0
numpy==1.26.4
python-multipart==0.0.9
